Design of the app was changhed. Some minor updates to all files

+++ app/src/main/java/uk/ac/ed/inf/coinz/MessageFragment.java
@@ -17,11 +17,9 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.QuerySnapshot;
@@ -34,7 +32,6 @@ import java.util.List;
 public class MessageFragment extends Fragment {
     private ListView listMessages;
     private TextView NoMessages;
-    private String cID;
     private List<DocumentSnapshot> list;
     private String tag = "MessageFragment";
     private ArrayList<String[]> coinsInformation;
@@ -42,7 +39,6 @@ public class MessageFragment extends Fragment {
     private FirebaseFirestore db;
     CollectionReference firestoreChat;
     private final String email = new CurrentUser().getEmail();
-    private ArrayList<String> coinsID = new ArrayList<>();
     private ArrayList<String> coins = new ArrayList<>();
     // Use com.google.firebase.Timestamp objects instead of java.util.Date objects
 
@@ -57,28 +53,25 @@ public class MessageFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        listMessages = (ListView) view.findViewById(R.id.ListMessages);
-        NoMessages = (TextView) view.findViewById(R.id.TextViewNoMessages);
+        listMessages = view.findViewById(R.id.ListMessages);
+        NoMessages = view.findViewById(R.id.TextViewNoMessages);
         NoMessages.setEnabled(false);
         NoMessages.setInputType(InputType.TYPE_NULL);
 
-        db = new MainActivity().db;
+        db = FirebaseFirestore.getInstance();
         coinsInformation = new ArrayList<>();
 
 
         db.collection("user:" + email)
-                .document("Coinz").collection("Chat").get().addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
-            @Override
-            public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
-                if (queryDocumentSnapshots.isEmpty()) {
+                .document("Coinz").collection("Chat").get()
+                .addOnSuccessListener((QuerySnapshot queryDocumentSnapshots) -> {
+                    if (queryDocumentSnapshots.isEmpty()) {
+                        //addToListView(null,false);
+                        //  listener.processResultFromFireStore(null,false);
+                        Log.d(tag, "No documents to show");
 
-                    //addToListView(null,false);
-                    //  listener.processResultFromFireStore(null,false);
-                    Log.d(tag, "No documents to show");
-
-                }
-            }
-        });
+                    }
+                });
 
         firestoreChat = db.collection("user:" + email)
                 .document("Coinz").collection("Chat");
@@ -86,43 +79,56 @@ public class MessageFragment extends Fragment {
     }
 
     private void realtimeUpdateListener() {
-        firestoreChat.addSnapshotListener(new EventListener<QuerySnapshot>() {
-            @Override
-            public void onEvent(@javax.annotation.Nullable QuerySnapshot queryDocumentSnapshots,
-                                @javax.annotation.Nullable FirebaseFirestoreException e) {
-                if (e != null) {
-                    System.err.println("Listen failed: " + e);
-                    return;
-                }
+        firestoreChat.addSnapshotListener(
+                (@javax.annotation.Nullable QuerySnapshot queryDocumentSnapshots,
+                 @javax.annotation.Nullable FirebaseFirestoreException e) -> {
+                    if (e != null) {
+                        System.err.println("Listen failed: " + e);
+                        return;
+                    }
+                    if (queryDocumentSnapshots != null) {
+//This is a reduntant step. It was done for not having warning only.
+                        list = queryDocumentSnapshots.getDocuments();
+                        addToListView(list);
+                    }
 
-                list = queryDocumentSnapshots.getDocuments();
-                addToListView(list, true);
-
-            }
-        });
+                });
 
 
     }
 
-    private void addToListView(List<DocumentSnapshot> l, boolean notnull) {
-        if (notnull) {
+    private void addToListView(List<DocumentSnapshot> l) {
+        if (l != null) {
             for (DocumentSnapshot d : l) {
-                String[] coinInf = new String[4];
-                String id = d.get("id").toString();
-                coinInf[0] = id;
-                double value = Double.parseDouble(d.get("value").toString());
-                coinInf[1] = value + "";
-                String currency = d.get("currency").toString();
-                coinInf[2] = currency;
-                Double exchangeRate = Double.parseDouble(d.get("exchangeRate").toString());
-                coinInf[3] = exchangeRate + "";
-                String sendBy = d.get("sendBy").toString();
+                Object idObj = d.get("id");
+                Object valueObj = d.get("value");
+                Object currencyObj = d.get("currency");
+                Object sendByObj = d.get("sendBy");
+                Object exchangeRateObj = d.get("exchangeRate");
 
-                double golds = value * exchangeRate;
-                coins.add(golds + " GOLDS you will get");
-                coinsID.add(id);
-                coinsInformation.add(coinInf);
+                if (idObj != null && valueObj != null && currencyObj != null &&
+                        sendByObj != null && exchangeRateObj != null) {
+                    String[] coinInf = new String[4];
+                    String id = idObj.toString();
+                    coinInf[0] = id;
+                    double value = Double.parseDouble(valueObj.toString());
+                    coinInf[1] = value + "";
+                    String currency = currencyObj.toString();
+                    coinInf[2] = currency;
+                    Double exchangeRate = Double.parseDouble(exchangeRateObj.toString());
+                    coinInf[3] = exchangeRate + "";
+                    String sendBy = sendByObj.toString();
+
+                    double golds = value * exchangeRate;
+                    coins.add(golds + " GOLDS you will get." + "Send by " + sendBy);
+                    coinsInformation.add(coinInf);
+                } else {
+                    Log.d(tag, "[addToListView] method " +
+                            "'get' produced NullPointException at d.get()");
+                }
             }
+        } else {
+            Log.d(tag, "[addToListView] List<DocumentSnapshot> l is empty");
         }
         setList();
 
@@ -130,23 +136,15 @@ public class MessageFragment extends Fragment {
 
     private void setList() {
         if (getContext() != null) {
-                ArrayAdapter adapter = new ArrayAdapter(requireContext(),
-                        android.R.layout.simple_list_item_1, coins);
-                listMessages.setAdapter(adapter);
-                listMessages.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-                    @Override
-                    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                        cID = coinsID.get(position);
-                        String coinInformation = coins.get(position);
-                        openDialog(position);
-
-
-                    }
-                });
-                if (!coins.isEmpty()) {
-                    NoMessages.setVisibility(View.GONE);
-                }
-            else{
+            ArrayAdapter adapter = new ArrayAdapter<>(requireContext(),
+                    android.R.layout.simple_list_item_1, coins);
+            listMessages.setAdapter(adapter);
+            listMessages.setOnItemClickListener((AdapterView<?> parent, View view, int position, long id) ->
+                    openDialog(position)
+            );
+            if (!coins.isEmpty()) {
+                NoMessages.setVisibility(View.GONE);
+            } else {
                 NoMessages.setVisibility(View.VISIBLE);
             }
         }
@@ -155,52 +153,53 @@ public class MessageFragment extends Fragment {
     private void openDialog(int index) {
         if (getActivity() != null) {
             AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-            builder.setMessage("Barev")
-                    .setTitle("Bank");
-            builder.setPositiveButton("Accept ", new DialogInterface.OnClickListener() {
-                public void onClick(DialogInterface dialog, int id) {
+            builder.setMessage("Message")
+                    .setTitle("Do you want to accept this coin ?");
 
-                    db.collection("user:" + email).document("Coinz").collection("Wallet")
-                            .whereEqualTo("id", id).get().addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
-                        @Override
-                        public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
-                            if (!queryDocumentSnapshots.isEmpty()) {
-                                Toast.makeText(getContext(),
-                                        "You have this coin in your wallet",
-                                        Toast.LENGTH_SHORT).show();
+            builder.setPositiveButton(("Accept"), (DialogInterface dialog, int id) -> {
+                String CoinID = coinsInformation.get(index)[0];
+                db.collection("user:" + email).document("Coinz").collection("Wallet")
+                        .whereEqualTo("id", CoinID).get().addOnSuccessListener((QuerySnapshot queryDocumentSnapshots) -> {
 
-                            } else {
-                                String coinInf[] = coinsInformation.get(index);
-                                HashMap<String, Object> coin = new HashMap<>();
-                                coin.put("id", coinInf[0]);
-                                coin.put("value", coinInf[1]);
-                                coin.put("currency", coinInf[2]);
-                                db.collection("user:" + email).document("Coinz")
-                                        .collection("Wallet").document(coinInf[0]).set(coin);
-                                firestoreChat.whereEqualTo("id", coinInf[0]).get().addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
-                                    @Override
-                                    public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
-                                        if (!queryDocumentSnapshots.isEmpty()) {
-                                            DocumentReference doc = queryDocumentSnapshots.
-                                                    getDocuments().get(0).getReference();
-                                            doc.delete();
-                                            coins = new ArrayList<>();
-                                            coinsID = new ArrayList<>();
-                                            coinsInformation = new ArrayList<>();
-                                        }
+                    if (!queryDocumentSnapshots.isEmpty()) {
+                        Toast.makeText(getContext(),
+                                "You have this coin in your wallet",
+                                Toast.LENGTH_SHORT).show();
+
+                    } else {
+
+                        String coinInf[] = coinsInformation.get(index);
+                        HashMap<String, Object> coin = new HashMap<>();
+                        coin.put("id", coinInf[0]);
+                        coin.put("value", coinInf[1]);
+                        coin.put("currency", coinInf[2]);
+
+                        db.collection("user:" + email).document("Coinz")
+                                .collection("Wallet").document(coinInf[0]).set(coin);
+                        db.collection("user:" + email).document("Coinz")
+                                .collection("NotCollected").document(coinInf[0]).delete();
+
+                        firestoreChat.whereEqualTo("id", coinInf[0]).get()
+                                .addOnSuccessListener((QuerySnapshot queryDS) -> {
+
+                                    if (!queryDS.isEmpty()) {
+                                        DocumentReference doc = queryDS.
+                                                getDocuments().get(0).getReference();
+                                        doc.delete();
+                                        coins = new ArrayList<>();
+                                        coinsInformation = new ArrayList<>();
                                     }
+
                                 });
-                            }
-                        }
-                    });
-                }
+                    }
+
+                });
             });
 
 
-            builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-                public void onClick(DialogInterface dialog, int id) {
-                }
-            });
+            builder.setNegativeButton(("Cancel"), (dialog, which) ->
+                    dialog.cancel()
+            );
             AlertDialog dialog = builder.create();
             dialog.show();
